+++ src/main/java/ru/yandex/money/gradle/plugins/library/LibraryProjectPlugin.java
@@ -4,6 +4,7 @@ import io.spring.gradle.dependencymanagement.DependencyManagementPlugin;
 import org.gradle.api.Plugin;
 import org.gradle.api.Project;
 import org.gradle.api.artifacts.dsl.RepositoryHandler;
+import org.gradle.api.plugins.JavaPlugin;
 import ru.yandex.money.gradle.plugins.library.changelog.CheckChangelogPlugin;
 import ru.yandex.money.gradle.plugins.library.dependencies.CheckDependenciesPlugin;
 import ru.yandex.money.gradle.plugins.library.git.BranchName;
@@ -29,6 +30,7 @@ public class LibraryProjectPlugin implements Plugin<Project> {
      * Все остальные настройки должны делаться в самом добавляемом плагине.
      */
     private static final Collection<Class<?>> PLUGINS_TO_APPLY = Arrays.asList(
+            JavaPlugin.class,
             DependencyManagementPlugin.class,
             ReadmePlugin.class,
             CheckChangelogPlugin.class,