+++ CHANGELOG.md
@@ -1,6 +1,10 @@
-# library-plugin
+# library-project-plugin
 
 ## NEXT_VERSION
+
+## [2.0.1]() (14-11-2018)
+
+Сборка при помощи gradle-project-plugin
 
 ## [2.0.0]() (03-11-2018)